attemp optimize

+++ common/gui/sprite.cpp
@@ -11,6 +11,6 @@ Sprite::Sprite()
 void DefS(const char* relative, Sprite* s, int offx, int offy)
 {
 	CreateTexture(s->texindex, relative, true, false);
-	s->offset[0] = offx;
-	s->offset[1] = offy;
+	s->offset[0] = (float)offx;
+	s->offset[1] = (float)offy;
 }
